// src/app/shared/model/vendor-contact-person-obj.model.ts
export class VendorContactPersonObj {
    VendorContactPersonId: number;
    VendorId: number;
    Name: string;
    MrEmployeePosition: string;
    Phone1: string;
    Phone2: string;
    Email: string;
    JoinDate: Date;
    IsOwner: boolean;
    Addr: string;
    Zipcode: string;
    SubZipcode: string;
    PhnArea1: string;
    PhnArea2: string;
    PhnExt1: string;
    PhnExt2: string;
    AreaCode1: string;
    AreaCode2: string;
    AreaCode3: string;
    AreaCode4: string;
    City: string;
    Province: string;
    RowVersion: string;
    constructor() { this.VendorContactPersonId = 0;}
}